Scripts/requirements.txt
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
boto3==1.26.106
botocore==1.29.106
colorama==0.4.6
comm==0.1.3
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
greenlet==2.0.2
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jmespath==1.0.1
jupyter_client==8.1.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.5
pycodestyle==2.10.0
Pygments==2.14.0
pyodbc==4.0.38
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.8
SQLAlchemy-Utils==0.40.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
